# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: sing-box-observability API
  version: 1.0.0
  description: >-
    sing-box-observability 使用的同源只读浏览器 API。
    这是 BFF 契约，不是上游 /observability/v1 API 的直接复制。
  license:
    name: GNU General Public License v3.0 or later
    identifier: GPL-3.0-or-later
...
servers:
  - url: /
security:
  - consoleSession: []
...
paths:
...
  /api/v1/session:
    get:
      tags: [会话]
      operationId: getSession
      security: []
      summary: 返回控制台认证是否启用以及当前是否已认证。
      responses:
        "200":
          description: 会话状态。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionInfo"
    post:
      tags: [会话]
      operationId: createSession
      security: []
      summary: 使用控制台令牌换取 HttpOnly Cookie。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: 已认证会话，并设置 sbox_observability_session。
          headers:
            Set-Cookie:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionInfo"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"
    delete:
      tags: [会话]
      operationId: deleteSession
      summary: 清除控制台会话 Cookie。
      responses:
        "204":
          description: 会话已清除。
  /api/v1/meta:
    get:
      tags: [可观测性]
      operationId: getMeta
      summary: 来源状态、capabilities 和采集器限制。
      responses:
        "200":
          description: 元数据和 capabilities。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MetaResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/overview:
    get:
      tags: [可观测性]
      operationId: getOverview
      summary: 当前快照、所选区间汇总、趋势点和简要排行。
      parameters:
        - $ref: "#/components/parameters/Range"
        - $ref: "#/components/parameters/From"
        - $ref: "#/components/parameters/To"
        - $ref: "#/components/parameters/Step"
      responses:
        "200":
          description: 概述数据；首次成功采集前 current 可以为 null。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OverviewResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /api/v1/rankings:
    get:
      tags: [可观测性]
      operationId: getRankings
      summary: 按一个受支持维度聚合已保留连接。
      parameters:
        - name: dimension
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/RankingDimension"
        - name: sort
          in: query
          schema:
            $ref: "#/components/schemas/RankingSort"
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - $ref: "#/components/parameters/Range"
        - $ref: "#/components/parameters/From"
        - $ref: "#/components/parameters/To"
      responses:
        "200":
          description: 根据保留连接生成排行，并合并查询时的活动连接。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RankingsResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: 请求的敏感维度已被上游关闭。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /api/v1/dimensions/series:
    get:
      tags: [可观测性]
      operationId: getDimensionSeries
      summary: 返回指定网络、入站或出站值的历史序列。
      parameters:
        - name: dimension
          in: query
          required: true
          schema:
            type: string
            enum: [network, inbound, outbound]
        - name: value
          in: query
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 512
        - $ref: "#/components/parameters/Range"
        - $ref: "#/components/parameters/From"
        - $ref: "#/components/parameters/To"
        - $ref: "#/components/parameters/Step"
      responses:
        "200":
          description: 维度序列；缺失时间桶的指标为 null。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DimensionSeriesResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /api/v1/connections/active:
    get:
      tags: [连接]
      operationId: getActiveConnections
      summary: 当前已校准的活动连接，按最新优先排列。
      parameters:
        - $ref: "#/components/parameters/Search"
        - $ref: "#/components/parameters/Network"
        - $ref: "#/components/parameters/Outbound"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: 活动连接分页。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectionPage"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/connections/recent:
    get:
      tags: [连接]
      operationId: getRecentConnections
      summary: 控制台保留的已关闭连接，按最新优先排列。
      parameters:
        - $ref: "#/components/parameters/Range"
        - $ref: "#/components/parameters/From"
        - $ref: "#/components/parameters/To"
        - $ref: "#/components/parameters/Search"
        - $ref: "#/components/parameters/Network"
        - $ref: "#/components/parameters/Outbound"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: 近期已关闭连接分页。
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConnectionPage"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /api/v1/events:
    get:
      tags: [可观测性]
      operationId: streamEvents
      summary: 用于来源状态和连接生命周期失效通知的下游 SSE。
      description: |
        使用 HttpOnly 会话 Cookie，并在有限内存回放窗口内支持 Last-Event-ID。
        事件名包括 `hello`、`source.state`、`connection.open`、
        `connection.close` 和 `resync`。keepalive 注释不是应用事件。
        收到 `resync` 或回放缺口后，客户端必须重新请求对应 REST 数据。
      parameters:
        - name: heartbeat
          in: query
          schema:
            type: string
            pattern: "^[0-9]+(s|m)$"
            default: 15s
        - name: lastEventId
          in: query
          description: 浏览器 EventSource 的回放回退参数；Last-Event-ID 请求头优先。
          schema:
            type: string
            pattern: "^[0-9]+$"
      responses:
        "200":
          description: 事件流；每个 data 行都是 EventEnvelope JSON 对象。
          content:
            text/event-stream:
              schema:
                type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
components:
  securitySchemes:
    consoleSession:
      type: apiKey
      in: cookie
      name: sbox_observability_session
  parameters:
    Range:
      name: range
      in: query
      description: 便捷时间范围，不能与 from/to 同时使用。
      schema:
        type: string
        enum: [15m, 1h, 6h, 24h, 7d, 30d, 90d]
        default: 1h
    From:
      name: from
      in: query
      description: 包含边界的 RFC3339 起点，必须与 to 成对且不能与 range 同时使用。
      schema:
        type: string
        format: date-time
    To:
      name: to
      in: query
      description: 不包含边界的 RFC3339 终点，必须与 from 成对且不能与 range 同时使用。
      schema:
        type: string
        format: date-time
    Step:
      name: step
      in: query
      description: 请求的时间桶大小；服务可以增大它以限制最多 720 个点。
      schema:
        type: string
        pattern: "^(auto|[0-9]+(s|m|h))$"
        default: auto
    Search:
      name: q
      in: query
      description: 在上游实际暴露的字段中进行不区分大小写搜索。
      schema:
        type: string
        maxLength: 200
    Network:
      name: network
      in: query
      schema:
        type: string
        maxLength: 32
    Outbound:
      name: outbound
      in: query
      schema:
        type: string
        maxLength: 512
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 500
        default: 50
    Offset:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
  responses:
    BadRequest:
      description: 请求格式错误。
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: 控制台会话缺失或无效。
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    UnprocessableEntity:
      description: 语法有效，但时间范围或筛选条件无法处理。
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    TooManyRequests:
      description: 已达到登录速率限制。
      headers:
        Retry-After:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    LoginRequest:
      type: object
      additionalProperties: false
      required: [token]
      properties:
        token:
          type: string
          minLength: 1
          maxLength: 4096
          writeOnly: true
    SessionInfo:
      type: object
      additionalProperties: false
      required: [authEnabled, authenticated]
      properties:
        authEnabled:
          type: boolean
        authenticated:
          type: boolean
    SourceState:
      type: string
      enum: [connecting, online, stale, offline, unauthorized]
    MetaResponse:
      type: object
      additionalProperties: false
      required: [apiVersion, appVersion, generatedAt, source, capabilities, collector]
      properties:
        apiVersion:
          type: string
          const: v1
        appVersion:
          type: string
        generatedAt:
          type: string
          format: date-time
        source:
          $ref: "#/components/schemas/SourceInfo"
        capabilities:
          oneOf:
            - $ref: "#/components/schemas/Capabilities"
            - type: "null"
        collector:
          $ref: "#/components/schemas/CollectorInfo"
...
    TimeWindow:
      type: object
      additionalProperties: false
      required: [from, to, stepSeconds]
      properties:
        from:
          type: string
          format: date-time
        to:
          type: string
          format: date-time
        stepSeconds:
          type: integer
          minimum: 1
...
    DimensionSeriesResponse:
      type: object
      additionalProperties: false
      required: [generatedAt, window, dimension, value, series]
      properties:
        generatedAt:
          type: string
          format: date-time
        window:
          $ref: "#/components/schemas/TimeWindow"
        dimension:
          type: string
          enum: [network, inbound, outbound]
        value:
          type: string
        series:
          type: array
          maxItems: 720
          items:
            $ref: "#/components/schemas/DimensionTimePoint"
...
    OverviewResponse:
      type: object
      additionalProperties: false
      required: [generatedAt, sourceState, window, current, rangeTotals, series, topOutbounds, urlTests, apiHealth]
      properties:
        generatedAt:
          type: string
          format: date-time
        sourceState:
          $ref: "#/components/schemas/SourceState"
        window:
          $ref: "#/components/schemas/TimeWindow"
        current:
          oneOf:
            - $ref: "#/components/schemas/StatusSnapshot"
            - type: "null"
        rangeTotals:
          oneOf:
            - $ref: "#/components/schemas/RangeTotals"
            - type: "null"
        series:
          type: array
          maxItems: 720
          items:
            $ref: "#/components/schemas/TimePoint"
        topOutbounds:
          type: array
          maxItems: 10
          items:
            $ref: "#/components/schemas/CompactRanking"
        urlTests:
          type: array
          maxItems: 1000
          items:
            $ref: "#/components/schemas/UrlTestResult"
        apiHealth:
          oneOf:
            - $ref: "#/components/schemas/APIHealthSnapshot"
            - type: "null"
    RankingDimension:
      type: string
      enum: [network, inbound, outbound, rule, domain, destination_ip, source, process, user]
    RankingSort:
      type: string
      enum: [traffic, connections, download, upload]
      default: traffic
...
    RankingsResponse:
      type: object
      additionalProperties: false
      required: [generatedAt, window, dimension, sort, total, data]
      properties:
        generatedAt:
          type: string
          format: date-time
        window:
          $ref: "#/components/schemas/TimeWindow"
        dimension:
          $ref: "#/components/schemas/RankingDimension"
        sort:
          $ref: "#/components/schemas/RankingSort"
        total:
          type: integer
          minimum: 0
        data:
          type: array
          maxItems: 100
          items:
            $ref: "#/components/schemas/RankingItem"
...
    ConnectionPage:
      type: object
      additionalProperties: false
      required: [generatedAt, total, limit, offset, data]
      properties:
        generatedAt:
          type: string
          format: date-time
        total:
          type: integer
          minimum: 0
        limit:
          type: integer
          minimum: 1
          maximum: 500
        offset:
          type: integer
          minimum: 0
        data:
          type: array
          items:
            $ref: "#/components/schemas/Connection"
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/ApiError"
...